+++ Makefile
@@ -8,19 +8,19 @@ GMOCK_DIR=googletest/googlemock
 
 # test locations
 SESSION_TEST=test/session_test.cpp session.cpp
-UTILS_TEST=test/utils_test.cpp 
+UTILS_TEST=test/utils_test.cpp
 SERVER_TEST=test/server_test.cpp
 PARSER_TEST=test/config_parser_test.cc
 HANDLER_TEST=test/handler_test.cpp
 
-all: server.o session.o main.o config_parser.o utils.o HttpRequest.o http_response.o handler.o
-	g++ -o web-server main.o server.o session.o config_parser.o utils.o HttpRequest.o http_response.o handler.o $(LDFLAGS) $(CXXFLAGS) 
+all: server.o session.o main.o config_parser.o utils.o http_request.o http_response.o handler.o
+	g++ -o web-server main.o server.o session.o config_parser.o utils.o http_request.o http_response.o handler.o $(LDFLAGS) $(CXXFLAGS)
 
 server.o: server.cpp server.h
 	g++ -c server.cpp $(LDFLAGS) $(CXXFLAGS)
 
 session.o: session.cpp session.h
-	g++ -c session.cpp $(LDFLAGS) $(CXXFLAGS) 
+	g++ -c session.cpp $(LDFLAGS) $(CXXFLAGS)
 
 config_parser.o: $(CP_LOC)config_parser.cc $(CP_LOC)config_parser.h
 	g++ -c $(CP_LOC)config_parser.cc -g $(CXXFLAGS)
@@ -31,8 +31,8 @@ utils.o: utils.cpp utils.h
 main.o: main.cpp server.h session.h $(CP_LOC)config_parser.h utils.h
 	g++ -c main.cpp $(LDFLAGS) $(CXXFLAGS)
 
-HttpRequest.o: HttpRequest.h HttpRequest.cpp
-	g++ -c HttpRequest.cpp $(LDFLAGS) $(CXXFLAGS)
+http_request.o: http_request.h http_request.cpp
+	g++ -c http_request.cpp $(LDFLAGS) $(CXXFLAGS)
 
 http_response.o: http_response.h http_response.cpp
 	g++ -c http_response.cpp $(LDFLAGS) $(CXXFLAGS)
@@ -42,14 +42,14 @@ handler.o: handler.h handler.cpp
 
 .PHONY: clean, all, test
 
-test: 
+test:
 	# Build web-server for integration test
-	make	
+	make
 	# Build gtest
 	g++ -std=c++0x -isystem ${GTEST_DIR}/include -I${GTEST_DIR} -pthread -c ${GTEST_DIR}/src/gtest-all.cc
 	ar -rv libgtest.a gtest-all.o
 	# Build tests
-	g++ -isystem ${GTEST_DIR}/include ${SESSION_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o HttpRequest.o http_response.o handler.o -o session_test ${LDFLAGS} ${CXXFLAGS}
+	g++ -isystem ${GTEST_DIR}/include ${SESSION_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o http_request.o http_response.o handler.o -o session_test ${LDFLAGS} ${CXXFLAGS}
 	g++ -isystem ${GTEST_DIR}/include ${UTILS_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o -o utils_test ${LDFLAGS} ${CXXFLAGS}
 	g++ -isystem ${GTEST_DIR}/include ${PARSER_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o -o config_parser_test ${LDFLAGS} ${CXXFLAGS}
 	g++ -isystem ${GTEST_DIR}/include ${SERVER_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o -o server_test ${LDFLAGS} ${CXXFLAGS}
